requirements.txt
# Runtime dependencies (mirrors [project.dependencies] in pyproject.toml)
grpcio>=1.62,<2
grpcio-tools>=1.62,<2
protobuf>=4.25,<7
cryptography>=42,<47

# Development / test dependencies
pytest>=8.0,<9
pytest-asyncio>=0.23
